backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pyjwt[crypto]>=2.9.0
httpx>=0.27.0
pyserial>=3.5
